- Added code highlight and refactored code
