+++ docs/src/components/Document/index.scss
@@ -31,24 +31,22 @@
         justify-content: space-between;
     }
 
-    code[data-lang] {
-        display: flex;
-        justify-content: space-between;
-        flex: 1;
-    }
-
     pre.highlight:hover {
         .copyButton {
             visibility: visible;
         }
 
-        code[data-lang]::after {
-            content: attr(data-lang);
-            text-transform: uppercase;
-            color: $darkgrey;
-            position: relative;
-            top: -9px;
+        .lang {
+            visibility: visible;
         }
+    }
+
+    .lang {
+        position: relative;
+        top: -14px;
+        text-transform: uppercase;
+        visibility: hidden;
+        color: $darkgrey;
     }
 
     #preview-in-playground {
@@ -111,7 +109,6 @@
       }
       
       .tooltip .tooltiptext {
-        // visibility: hidden;
         width: 120px;
         background-color: $gray-80;
         color: white;
@@ -122,9 +119,96 @@
         /* Position the tooltip */
         position: absolute;
         right: -35px;
-        top: 15px;
+        top: 10px;
         z-index: 1;
       }
+
+    /*! Adapted from the GitHub style by Vasily Polovnyov <vast@whiteants.net> */
+    .hljs-comment,
+    .hljs-quote {
+        color: #998;
+        font-style: italic;
+    }
+
+    .hljs-keyword,
+    .hljs-selector-tag,
+    .hljs-subst {
+        color: #333;
+        font-weight: $font-weight-bold;
+    }
+
+    .hljs-number,
+    .hljs-literal,
+    .hljs-variable,
+    .hljs-template-variable,
+    .hljs-tag .hljs-attr {
+        color: #008080;
+    }
+
+    .hljs-string,
+    .hljs-doctag {
+        color: #d14;
+    }
+
+    .hljs-title,
+    .hljs-section,
+    .hljs-selector-id {
+        color: #900;
+        font-weight: $font-weight-bold;
+    }
+
+    .hljs-subst {
+        font-weight: normal;
+    }
+
+    .hljs-type,
+    .hljs-class .hljs-title {
+        color: #458;
+        font-weight: $font-weight-bold;
+    }
+
+    .hljs-tag,
+    .hljs-name,
+    .hljs-attribute {
+        color: #000080;
+        font-weight: normal;
+    }
+
+    .hljs-regexp,
+    .hljs-link {
+        color: #009926;
+    }
+
+    .hljs-symbol,
+    .hljs-bullet {
+        color: #990073;
+    }
+
+    .hljs-built_in,
+    .hljs-builtin-name {
+        color: #0086b3;
+    }
+
+    .hljs-meta {
+        color: #999;
+        font-weight: $font-weight-bold;
+    }
+
+    .hljs-deletion {
+        background: #fdd;
+    }
+
+    .hljs-addition {
+        background: #dfd;
+    }
+
+    .hljs-emphasis {
+        font-style: italic;
+    }
+
+    .hljs-strong {
+        font-weight: var(--monospace-font-weight-bold);
+    }
 
 }
