+++ docs/src/components/Document/index.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import './index.scss';
-import Copy from '../../assets/svg/copy.svg';
-import addCopyFeature from './copyFunctionality';
+import { customizeDocContent } from './helper';
 
 const Document = (props: { docTitle: string; docContent: string }) => {
 
@@ -10,28 +9,8 @@ const Document = (props: { docTitle: string; docContent: string }) => {
     useEffect(() => {
         const divElement = document.createElement('div');
         divElement.innerHTML = props.docContent;
-        addCopyButton(divElement);
+        customizeDocContent(divElement, setDocContent);
     }, [props.docContent]);
-
-    const addCopyButton = (divElement: HTMLElement) => {
-        const allTags = divElement.querySelectorAll('.listingblock>.content>.highlight>code');
-        if (allTags.length > 0) {
-            divElement.querySelectorAll('.listingblock>.content>.highlight>code').forEach((tag) => {
-                const buttonElement = document.createElement('button');
-                buttonElement.setAttribute('class', 'copyButton');
-                const imageElement = document.createElement('img');
-                imageElement.src = Copy;
-                imageElement.alt = 'Copy';
-                imageElement.title = 'Copy';
-                buttonElement.appendChild(imageElement);
-                tag.parentElement.appendChild(buttonElement);
-            });
-        }
-        setDocContent(divElement.innerHTML)
-        setTimeout(() => {
-            addCopyFeature();
-        }, 500);
-    };
 
     return (
         <div className="documentWrapper">
